moved InMemory code from API to Persistence.InMemory project
