+++ Fabric.Authorization.API/Extensions/TinyIocExtensions.cs
@@ -7,8 +7,8 @@ using Fabric.Authorization.Domain.Resolvers.Permissions;
 using Fabric.Authorization.Domain.Services;
 using Fabric.Authorization.Domain.Stores;
 using Fabric.Authorization.Domain.Stores.CouchDB;
-using Fabric.Authorization.Domain.Stores.InMemory;
 using Fabric.Authorization.Domain.Stores.Services;
+using Fabric.Authorization.Persistence.InMemory.Stores;
 using Fabric.Authorization.Persistence.SqlServer.Services;
 using Fabric.Authorization.Persistence.SqlServer.Stores;
 using Nancy.TinyIoc;
