--- Fabric.Authorization.Domain/Stores/InMemory/InMemoryDbBootstrapper.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Fabric.Authorization.Domain.Stores.InMemory
-{
-    public class InMemoryDbBootstrapper : IDbBootstrapper
-    {
-        public void Setup()
-        {
-            return;
-        }
-    }
-}
