+++ Fabric.Authorization.IntegrationTests/IntegrationTestsFixture.cs
@@ -103,8 +103,8 @@ namespace Fabric.Authorization.IntegrationTests
             CouchDbSettings config = new CouchDbSettings()
             {
                 DatabaseName = "integration-" + DateTime.UtcNow.Ticks,
-                Username = "",
-                Password = "",
+                Username = "admin",
+                Password = "admin",
                 Server = "http://127.0.0.1:5984"
             };
