+++ Fabric.Authorization.Persistence.InMemory/Fabric.Authorization.Persistence.InMemory.csproj
@@ -6,8 +6,6 @@
 
   <ItemGroup>
     <Folder Include="DependencyInjection\" />
-    <Folder Include="Services\" />
-    <Folder Include="Stores\" />
   </ItemGroup>
 
   <ItemGroup>
