+++ Fabric.Authorization.Persistence.InMemory/Services/InMemoryDbBootstrapper.cs
@@ -1,0 +1,11 @@
+﻿using Fabric.Authorization.Domain.Stores;
+
+namespace Fabric.Authorization.Persistence.InMemory.Services
+{
+    public class InMemoryDbBootstrapper : IDbBootstrapper
+    {
+        public void Setup()
+        {
+        }
+    }
+}
