+++ Fabric.Authorization.Persistence.InMemory/Stores/InMemoryGroupStore.cs
@@ -3,9 +3,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Fabric.Authorization.Domain.Exceptions;
 using Fabric.Authorization.Domain.Models;
-using Fabric.Authorization.Domain.Stores.Services;
+using Fabric.Authorization.Domain.Stores;
 
-namespace Fabric.Authorization.Domain.Stores.InMemory
+namespace Fabric.Authorization.Persistence.InMemory.Stores
 {
     public class InMemoryGroupStore : InMemoryFormattableIdentifierStore<Group>, IGroupStore
     {
