+++ Fabric.Authorization.Persistence.InMemory/Stores/InMemoryPermissionStore.cs
@@ -5,8 +5,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Fabric.Authorization.Domain.Exceptions;
 using Fabric.Authorization.Domain.Models;
+using Fabric.Authorization.Domain.Stores;
 
-namespace Fabric.Authorization.Domain.Stores.InMemory
+namespace Fabric.Authorization.Persistence.InMemory.Stores
 {
     public class InMemoryPermissionStore : InMemoryFormattableIdentifierStore<Permission>, IPermissionStore
     {
