+++ Fabric.Authorization.UnitTests/Fabric.Authorization.UnitTests.csproj
@@ -32,6 +32,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Fabric.Authorization.API\Fabric.Authorization.API.csproj" />
     <ProjectReference Include="..\Fabric.Authorization.Domain\Fabric.Authorization.Domain.csproj" />
+    <ProjectReference Include="..\Fabric.Authorization.Persistence.InMemory\Fabric.Authorization.Persistence.InMemory.csproj" />
   </ItemGroup>
 
   <ItemGroup>
