+++ Fabric.Authorization.UnitTests/RequestHooks/RequestHooksTests.cs
@@ -3,8 +3,8 @@ using System.Security.Claims;
 using Fabric.Authorization.API.Constants;
 using Fabric.Authorization.API.Models;
 using Fabric.Authorization.API.Modules;
-using Fabric.Authorization.Domain.Stores.InMemory;
 using Fabric.Authorization.Domain.Stores.Services;
+using Fabric.Authorization.Persistence.InMemory.Stores;
 using Moq;
 using Nancy;
 using Nancy.Testing;
